django/requirements.txt
Django
future
factory-boy
geonamescache
gunicorn
psycopg2-binary
cryptography
celery
googlemaps
requests
Pillow
sorl-thumbnail

django-admin-list-filter-dropdown
django-allauth
# django-anymail[sendgrid]
django-dotenv
django-extensions
django-filter
django-modeltranslation
django-admin-sortable2
django-gfklookupwidget
dj-rest-auth[with_social]==4.0.0
django-celery-email
django-redis
django-rosetta
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-admin-rangefilter
django-filter
django-nonrelated-inlines
django-import-export
django-ordered-model
django-simplefeedback==1.2.2
django-simple-history
django-ckeditor
django-taggit
django-ckeditor
coreapi
drf-yasg
hashids

pytest
pytest-cov
pytest-django
pytest-testmon
pytest-watch
pytest-xdist
pytest-timeout
ipdb
flake8
dj-inmemorystorage

pycountry
hashids
sentry-sdk

redis
kombu
raven
